custom-elements/registering-custom-elements/definition-construction-algorithm-test-006.html

<!DOCTYPE html>
<html>
<head>
<title>If namespace is SVG namespace and name is null then error must be thrown</title>
<meta name="author" title="Vasiliy Degtyarev" href="mailto:vasya@unipro.ru">
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If namespace is SVG namespace and name is null then error must be thrown">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var svg = doc.createElementNS(SVG_NAMESPACE, 'xsvg');
    var proto = Object.create(svg.constructor.prototype);
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function () {
        doc.registerElement('x-svg-1', {prototype: proto});
    }, 'Exception should be thrown in case of attempt to register ' + 
    'a custom element with SVG namespace and name is not specified');
    
}, 'Error should be thrown if namespace is SVG and local name is not specified');


test(function() {
    var doc = newHTMLDocument();
    var svg = doc.createElementNS(SVG_NAMESPACE, 'xsvg');
    var proto = Object.create(svg.constructor.prototype);
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function () {
        doc.registerElement('x-svg-2', {prototype: proto, extends: null});
    }, 'Exception should be thrown in case of attempt to register ' + 
    'a custom element with SVG namespace and name is null');

}, 'Error should be thrown if namespace is SVG and local name is null');

</script>
</body>
</html>
